Fixes a base directory error that occurs when executing the script from outside hoststatus/
The script should be able to execute from anywhere on disk. This was resolved by setting the correct base directory and then using absolute paths. This could possibly have been remedied by cd-ing to the base directory in the wrapper script and then cd-ing to the original directory upon exit.